chore: Scala 3 leftover (#367)

+++ akka-projection-r2dbc/src/test/scala/akka/projection/r2dbc/R2dbcOffsetStoreSpec.scala
@@ -53,7 +53,7 @@ class R2dbcOffsetStoreSpec
             .createStatement(selectLastSql)
             .bind(0, projectionId.name)
             .bind(1, projectionId.key),
-        row => Instant.ofEpochMilli(row.get("last_updated", classOf[java.lang.Long])))
+        row => Instant.ofEpochMilli(row.get("last_updated", classOf[java.lang.Long]).longValue()))
       .futureValue
       .getOrElse(throw new RuntimeException(s"no records found for $projectionId"))
   }
